+++ tests/testthat/test-utilities-scenarios.R
@@ -3,15 +3,27 @@
 projectConfiguration <- createDefaultProjectConfiguration(test_ProjectConfiguration())
 defaultOutputPath <- "Organism|PeripheralVenousBlood|Aciclovir|Plasma (Peripheral Venous Blood)"
 
-test_that("It runs one scenario without specifying output paths", {
+test_that("It stops with an error if the excel file defines a parameter that is
+          not present", {
   # Define which scenarios to run
-  scenarioNames <- c("TestScenario")
+  scenarioNames <- c("TestScenario_missingParam")
   # Create `ScenarioConfiguration` objects from excel files
   scenarioConfigurations <- readScenarioConfigurationFromExcel(
     scenarioNames = scenarioNames,
     projectConfiguration = projectConfiguration
   )
-  scenarios <- createScenarios(scenarioConfigurations = scenarioConfigurations)
+  expect_error(createScenarios(scenarioConfigurations = scenarioConfigurations, stopIfParameterNotFound = TRUE))
+})
+
+test_that("It runs one scenario without specifying output paths", {
+  # Define which scenarios to run
+  scenarioNames <- c("TestScenario_missingParam")
+  # Create `ScenarioConfiguration` objects from excel files
+  scenarioConfigurations <- readScenarioConfigurationFromExcel(
+    scenarioNames = scenarioNames,
+    projectConfiguration = projectConfiguration
+  )
+  scenarios <- createScenarios(scenarioConfigurations = scenarioConfigurations, stopIfParameterNotFound = FALSE)
 
   simulatedScenarios <- runScenarios(
     scenarios = scenarios